# apps/command/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "additionalProperties": false,
        "description": "``code`` is machine-readable; ``message`` says what happened and what to do.",
        "properties": {
          "code": {
            "description": "e.g. run_not_found",
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "description": "e.g. 'Run MUM-... is not baked yet. Press Play on the replay or Compute live.'",
            "title": "Message",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Every error response: ``{\"error\": {\"code\", \"message\", \"run_id\"}}``.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "Feature": {
        "additionalProperties": false,
        "properties": {
          "geometry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Point"
              },
              {
                "$ref": "#/components/schemas/LineString"
              },
              {
                "$ref": "#/components/schemas/MultiLineString"
              },
              {
                "$ref": "#/components/schemas/Polygon"
              },
              {
                "$ref": "#/components/schemas/MultiPolygon"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          },
          "id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "type": {
            "const": "Feature",
            "default": "Feature",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "geometry"
        ],
        "title": "Feature",
        "type": "object"
      },
      "FeatureCollection": {
        "additionalProperties": false,
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "maxItems": 4,
                "minItems": 4,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "title": "Features",
            "type": "array"
          },
          "type": {
            "const": "FeatureCollection",
            "default": "FeatureCollection",
            "title": "Type",
            "type": "string"
          }
        },
        "title": "FeatureCollection",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LineString": {
        "additionalProperties": false,
        "properties": {
          "coordinates": {
            "items": {
              "anyOf": [
                {
                  "maxItems": 2,
                  "minItems": 2,
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array"
                },
                {
                  "maxItems": 3,
                  "minItems": 3,
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array"
                }
              ]
            },
            "minItems": 2,
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "LineString",
            "default": "LineString",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "LineString",
        "type": "object"
      },
      "MultiLineString": {
        "additionalProperties": false,
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "maxItems": 2,
                    "minItems": 2,
                    "prefixItems": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "number"
                      }
                    ],
                    "type": "array"
                  },
                  {
                    "maxItems": 3,
                    "minItems": 3,
                    "prefixItems": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "number"
                      },
                      {
                        "type": "number"
                      }
                    ],
                    "type": "array"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "MultiLineString",
            "default": "MultiLineString",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "MultiLineString",
        "type": "object"
      },
      "MultiPolygon": {
        "additionalProperties": false,
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "items": {
                  "anyOf": [
                    {
                      "maxItems": 2,
                      "minItems": 2,
                      "prefixItems": [
                        {
                          "type": "number"
                        },
                        {
                          "type": "number"
                        }
                      ],
                      "type": "array"
                    },
                    {
                      "maxItems": 3,
                      "minItems": 3,
                      "prefixItems": [
                        {
                          "type": "number"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "number"
                        }
                      ],
                      "type": "array"
                    }
                  ]
                },
                "type": "array"
              },
              "type": "array"
            },
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "MultiPolygon",
            "default": "MultiPolygon",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "MultiPolygon",
        "type": "object"
      },
...
      "Point": {
        "additionalProperties": false,
        "properties": {
          "coordinates": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "maxItems": 3,
                "minItems": 3,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              }
            ],
            "title": "Coordinates"
          },
          "type": {
            "const": "Point",
            "default": "Point",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "Point",
        "type": "object"
      },
      "Polygon": {
        "additionalProperties": false,
        "description": "Rings of positions; the first ring is the exterior. Rings must close (first == last).",
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "maxItems": 2,
                    "minItems": 2,
                    "prefixItems": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "number"
                      }
                    ],
                    "type": "array"
                  },
                  {
                    "maxItems": 3,
                    "minItems": 3,
                    "prefixItems": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "number"
                      },
                      {
                        "type": "number"
                      }
                    ],
                    "type": "array"
                  }
                ]
              },
              "type": "array"
            },
            "minItems": 1,
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "Polygon",
            "default": "Polygon",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "Polygon",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Team VIT, SIH 2026 PS SIH26085"
    },
    "description": "Street-level urban flood nowcasting digital twin (SIH 2026, PS SIH26085, MoES). Every response carries `run_id` and `valid_ts` where a run is involved; errors use `{ error: { code, message, run_id } }`. Endpoints answering 501 are contract-only until their engine's phase lands.",
    "title": "VARUNA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/city/{city}/layers/{name}": {
      "get": {
        "description": "One static city layer as GeoJSON, straight from ``city/<city>/map/``.",
        "operationId": "city_layer_v1_city__city__layers__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "schema": {
              "title": "City",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "enum": [
                "segments",
                "drains",
                "drain_nodes",
                "assets",
                "hotspots",
                "buildings",
                "units",
                "depressions"
              ],
              "title": "Name",
              "type": "string"
            }
          },
          {
            "description": "minlon,minlat,maxlon,maxlat (WGS84); omit for the AOI",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "minlon,minlat,maxlon,maxlat (WGS84); omit for the AOI",
              "title": "Bbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureCollection"
                }
              }
            },
            "description": "GeoJSON FeatureCollection in WGS84, simplified with a 2 m tolerance"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "City or layer not built yet"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Static layer simplified for the map (segments, drains, assets, hotspots, buildings)",
        "tags": [
          "city"
        ]
      }
    },
...
    "/v1/onboard/city/{city}": {
      "get": {
        "description": "What the wizard asks on load, so a reopened tab rejoins a build already running.\n\nWith no job in this process the answer is ``status: \"none\"`` plus ``previous``: the city's\nlast good build as `city/<city>/onboard_last.json` recorded it (its log lines, each with the\ninstant it was captured, its per-step milliseconds and its first run), or null when no build\nwas ever recorded. ``last_attempt`` is set when a later build did not finish, so the wizard\ncan say that it failed without losing the forecast the good one made.",
        "operationId": "onboard_latest_v1_onboard_city__city__get",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "schema": {
              "title": "City",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Onboard Latest V1 Onboard City  City  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The most recent job for a city",
        "tags": [
          "onboard"
        ]
      }
    },
...
    "/v1/route": {
      "post": {
        "description": "Plan one trip.\n\nBody: ``{origin, destination, depart_at?, profile?, risk_tolerance?, run_id?, spread?,\ntrip_id?, explain?}``.\n\nReturns both routes, because the comparison is the product: a dispatcher who only sees the\nsafe route has no way to judge whether the detour was worth it.\n\n``spread`` (default true) adds ``corridors[]`` - up to three safe roads with the share of\ntraffic the policy gives each, and the one this request is assigned to. ``trip_id`` makes\nthat assignment stable for a trip, so a reader who reloads is not sent somewhere else; it is\nthe client's own random id and nothing is stored against it. ``explain`` (default true) adds\n``reasons[]``, which is structured data - the frontend writes the sentences (TECH_SPEC 3.2).",
        "operationId": "route_v1_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Route V1 Route Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route around the forecast water, beside what a naive router does",
        "tags": [
          "route"
        ]
      }
    },
...
  },
...
}